fix duplicate type file
